HSEARCH-3554 Move query-related ITs to a single package
