+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/testsupport/stub/StubDocumentReferenceTransformer.java
@@ -4,13 +4,15 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.backend.tck.search;
+package org.hibernate.search.integrationtest.backend.tck.testsupport.stub;
 
-import org.hibernate.search.engine.search.loading.spi.ObjectLoader;
+import java.util.function.Function;
+
+import org.hibernate.search.engine.search.DocumentReference;
 
 /**
  * The only purpose of this class is to avoid unchecked cast warnings when creating mocks.
  */
-public interface StubObjectLoader
-		extends ObjectLoader<StubTransformedReference, StubLoadedObject> {
+public interface StubDocumentReferenceTransformer
+		extends Function<DocumentReference, StubTransformedReference> {
 }
