+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/testsupport/stub/StubTransformedHit.java
@@ -4,14 +4,14 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.backend.tck.search;
+package org.hibernate.search.integrationtest.backend.tck.testsupport.stub;
 
 import org.hibernate.search.engine.search.DocumentReference;
 
-public class StubLoadedObject {
+public class StubTransformedHit {
 	private final DocumentReference documentReference;
 
-	public StubLoadedObject(DocumentReference documentReference) {
+	public StubTransformedHit(DocumentReference documentReference) {
 		this.documentReference = documentReference;
 	}
